# one-ui/ui/api/swagger/mocks.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample of a DataHub Swagger document defining endpoints to be mocked",
    "version": "1.0.0",
    "title": "Data Hub Curation API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:4200",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/map-search": {
      "post": {
        "tags": [
          "mapping search"
        ],
        "summary": "Search documents",
        "description": "....",
        "operationId": "search",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/AnyObject"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "x-swagger-router-controller": "Search"
      }
    },
    "/map-search/sjsSearch": {
      "post": {
        "tags": [
          "mapping search"
        ],
        "summary": "Search documents",
        "description": "....",
        "operationId": "sjsSearch",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/AnyObject"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "x-swagger-router-controller": "Search"
      }
    },
    "/map-search/doc": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get a document by database and URI",
        "description": "....",
        "operationId": "getDocumentByDatabaseAndURI",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "database",
            "in": "query",
            "type": "string"
          },
          {
            "name": "docUri",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/AnyObject"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "x-swagger-router-controller": "Documents"
      }
    },
...
    "/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get model information",
        "description": "....",
        "operationId": "getModel",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/modelInfo"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "x-swagger-router-controller": "models"
      }
    },
    "/models/{modelName}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get model information",
        "description": "....",
        "operationId": "getModelbyName",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "modelName",
            "type": "string",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/modelInfo"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "x-swagger-router-controller": "models"
      }
    },
    "/models/job-info": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Returns job info for all models",
        "description": "Returns job info for all models",
        "operationId": "getJobInfoForAllModels",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/jobInfoForModel"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "x-swagger-router-controller": "Jobs"
      }
    },
    "/models/{modelName}/job-info": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Returns job info for given model",
        "description": "Returns job info for given model",
        "operationId": "getJobInfoForModel",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "modelName",
            "type": "string",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/jobInfoForModel"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "x-swagger-router-controller": "Jobs"
      }
    },
...
  },
  "definitions": {
    "AnyObject": {
      "type": "object",
      "description": "This is a temporary definition for any object until details can be fleshed out",
      "properties": {}
    },
...
    "jobInfoForModel": {
      "type": "object",
      "description": "Installation details about a project",
      "example": {
        "entityCollection": "Order",
        "latestJobId": "75162145-665b-444e-bf6c-31d110abb8da",
        "latestJobDateTime": "2020-03-05T13:06:52.00875-08:00"
      }
    },
    "modelInfo": {
      "type": "object",
      "description": "Installation details about a project",
      "example": {
        "info": {
          "title": "FirstName",
          "version": "0.0.1",
          "baseUri": "http://example.org/"
        },
        "definitions": {
          "FirstName": {
            "required": [
            ],
            "pii": [
            ],
            "elementRangeIndex": [
            ],
            "rangeIndex": [
            ],
            "wordLexicon": [
            ],
            "properties": {
              "value": {
                "datatype": "string",
                "collation": "http://marklogic.com/collation/codepoint"
              },
              "prefix": {
                "datatype": "string",
                "collation": "http://marklogic.com/collation/codepoint"
              }
            }
          }
        }
      }
    },
...
    "error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  },
...
}